+++ src/main/obsEnums.ts
@@ -116,3 +116,53 @@ export enum ESupportedEncoders {
   JIM_NVENC = 'jim_nvenc',
   OBS_X264 = 'obs_x264',
 }
+
+export const enum ERangeType {
+  Default,
+  Partial,
+  Full,
+}
+
+export const enum EVideoFormat {
+  None,
+  I420,
+  NV12,
+  YVYU,
+  YUY2,
+  UYVY,
+  RGBA,
+  BGRA,
+  BGRX,
+  Y800,
+  I444,
+  BGR3,
+  I422,
+  I40A,
+  I42A,
+  YUVA,
+  AYUV,
+}
+
+export const enum EColorSpace {
+  Default,
+  CS601,
+  CS709,
+  CSSRGB,
+  CS2100PQ,
+  CS2100HLG,
+}
+
+export const enum EScaleType {
+  Disable,
+  Point,
+  Bicubic,
+  Bilinear,
+  Lanczos,
+  Area,
+}
+
+export const enum EFPSType {
+  Common,
+  Integer,
+  Fractional,
+}