+++ src/renderer/App.tsx
@@ -79,6 +79,7 @@ const Application = () => {
         return {
           ...prevState,
           videoFullScreen: false,
+          videoFilterQuery: '',
         };
       });
     });