+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Issue 446](https://github.com/aza547/wow-recorder/issues/446) - Fix to darkness in AMD encodings, and improvements for other encoders.
 - Fixed an issue with some encounters not working as search bar queries.
 - Hide the "Unknown Raid" text to make app maintenance easier.
+- Fix a bug where a search query would not be forgotten when a new video is saved.
 
 ## [3.19.0] - 2023-10-03
 ### Added